+++ source/extensions/omni.flux.stage_manager.factory/docs/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.1.1]
+### Changed
+- Added the ability to set the context from the core to the interaction plugin
+
+### Fixed
+- Fixed model serialization
+
 ## [1.1.0]
 ### Changed
 - Implemented plugin base classes