+++ source/extensions/omni.flux.stage_manager.plugin.interaction.usd/docs/CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.1.1]
+### Added
+- Added `context_name` field to the USD base
+
+### Changed
+- Propagate the `context_name` field for all children plugins
+
+### Fixed
+- Fixed dependencies
+
 ## [1.1.0]
 ### Changed
 - Added more compatible plugins for `AllPrimsInteractionPlugin`